Add filter context to validatePayload. Refactor to solve circular dependancy.
