+++ api/tests/services/authorization.test.ts
@@ -89,8 +89,8 @@ describe('Integration Tests', () => {
 					{ id: '6107c897-9182-40f7-b22e-4f044d1258d2', title: 'Hello 2', publish_date: nowTZ },
 				];
 
-				it.each(newPosts)('presets are added correctly to the payload', (payload) => {
-					const payloadWithPresets = service.validatePayload('create', tableName, payload);
+				it.each(newPosts)('presets are added correctly to the payload', async (payload) => {
+					const payloadWithPresets = await service.validatePayload('create', tableName, payload);
 
 					expect(payloadWithPresets).toHaveProperty('publish_date');
 					expect(payloadWithPresets.publish_date).not.toEqual('$NOW');
@@ -101,20 +101,18 @@ describe('Integration Tests', () => {
 			describe('update', () => {
 				const nowTZ = new Date().toISOString();
 				const newPosts = [
-					{ id: 'd66ec139-2655-48c1-9d9a-4753f98a9ee7', title: 'Hello World' },
-					{ id: '6107c897-9182-40f7-b22e-4f044d1258d2', title: 'Hello World 2', publish_date: nowTZ },
+					{ id: 'd66ec139-2655-48c1-9d9a-4753f98a9ee7', title: 'Should succeed' },
+					{ id: '6107c897-9182-40f7-b22e-4f044d1258d2', title: 'Should fail', publish_date: nowTZ },
 				];
 
-				it.each(newPosts)('validates the payload correctly', (payload) => {
+				it.each(newPosts)('validates the payload correctly', async (payload) => {
 					// Should only validate if field is set in the payload
 					if (!payload.publish_date) {
-						const payloadWithPresets = service.validatePayload('update', tableName, payload);
+						const payloadWithPresets = await service.validatePayload('update', tableName, payload);
 
 						expect(payloadWithPresets).not.toHaveProperty('publish_date');
 					} else {
-						expect(() => {
-							service.validatePayload('update', tableName, payload);
-						}).toThrow(Array);
+						expect(() => service.validatePayload('update', tableName, payload)).rejects.toBeInstanceOf(Array);
 					}
 				});
 			});
